\name{California_hosp_linelist}
\docType{data}
\alias{California_hosp_linelist}
\title{Simulated hospitalization linelist in California}
\description{
  This line list reports the hospitalizations of a synthetic epidemic in California.
}
\usage{data("California_hosp_linelist")}
\format{A data frame containing 7 columns and 176,033 observations}
\source{GLEAM Model}
\references{
  https://covid19.gleamproject.org/mobility
}
\keyword{datasets}
