\name{California_infections}
\docType{data}
\alias{California_infections}
\title{Simulated Infections in California}
\description{
  Time series of the SLIR model simulation of a synthetic epidemic across California.
}
\usage{data("California_infections")}
\format{A data frame containing 9 columns and 104,980 observations}
\source{GLEAM Model}
\references{
  https://covid19.gleamproject.org/mobility
}
\keyword{datasets}
