\name{Massachusetts_hosp_linelist}
\docType{data}
\alias{Massachusetts_hosp_linelist}
\title{Simulated hospitalization linelist in Massachusetts}
\description{
  This line list reports the hospitalizations of a synthetic epidemic in Massachusetts.
}
\usage{data("Massachusetts_hosp_linelist")}
\format{A data frame containing 7 columns and 31,157 observations}
\source{GLEAM Model}
\references{
  https://covid19.gleamproject.org/mobility
}
\keyword{datasets}
