\name{Massachusetts_infections}
\docType{data}
\alias{Massachusetts_infections}
\title{Simulated Infections in Massachusetts}
\description{
  Time series of the SLIR model simulation of a synthetic epidemic across Massachusetts.
}
\usage{data("Massachusetts_infections")}
\format{A data frame containing 9 columns and 25,340 observations}
\source{GLEAM Model}
\references{
  https://covid19.gleamproject.org/mobility
}
\keyword{datasets}
