\name{North_Carolina_hosp_linelist}
\docType{data}
\alias{North_Carolina_hosp_linelist}
\title{Simulated hospitalization linelist in North Carolina}
\description{
  This line list reports the hospitalizations of a synthetic epidemic in North Carolina.
}
\usage{data("North_Carolina_hosp_linelist")}
\format{A data frame containing 7 columns and 45,974 observations}
\source{GLEAM Model}
\references{
  https://covid19.gleamproject.org/mobility
}
\keyword{datasets}
