\name{North_Carolina_infections}
\docType{data}
\alias{North_Carolina_infections}
\title{Simulated Infections in North Carolina}
\description{
  Time series of the SLIR model simulation of a synthetic epidemic across North Carolina.
}
\usage{data("North_Carolina_infections")}
\format{A data frame containing 9 columns and 181,000 observations}
\source{GLEAM Model}
\references{
  https://covid19.gleamproject.org/mobility
}
\keyword{datasets}
