\name{Weekly_US_Deaths_20240924}
\docType{data}
\alias{Weekly_US_Deaths_20240924}
\title{Unites States COVID-19 Mortality Data}
\description{
  Weekly US COVID-19 Mortality data, pulled on 2024-09-24
}
\usage{data("Weekly_US_Deaths_20240924")}
\format{A data frame containing 4 columns and 245 observations}
\source{CDC}
\references{
  https://covid.cdc.gov/covid-data-tracker/#maps_percent-covid-deaths
}
\keyword{datasets}
