\name{Weekly_US_Hospitalizations_20240925}
\docType{data}
\alias{Weekly_US_Hospitalizations_20240925}
\title{Unites States COVID-19 Hospital Admissions Data}
\description{
  Weekly US COVID-19 Mortality data, pulled on 2024-09-25
}
\usage{data("Weekly_US_Hospitalizations_20240925")}
\format{A data frame containing 64 columns and 12,255 observations}
\source{CDC}
\references{
  https://data.cdc.gov/Public-Health-Surveillance/Weekly-United-States-Hospitalization-Metrics-by-Ju/aemt-mg7g/about_data
}
\keyword{datasets}
